Fix sendUseItemOnNpc Integer unbox; enforce single protect prayer
- Unbox pickedEntityId (Integer → int) at the sendUseItemOnNpc call site — fixes CI compilation error
- When activating a protect prayer, deactivate all other protect prayers first so only one overhead protection can be active at a time

+++ client/src/main/java/com/osrs/client/GameScreen.java
@@ -4602,7 +4602,7 @@ public class GameScreen extends ApplicationAdapter {
                         && leftClickH != null
                         && !leftClickH.isPlayer(pickedEntityId)) {
                     if (nettyClient != null) {
-                        nettyClient.sendUseItemOnNpc(selectedInventorySlot, pickedEntityId);
+                        nettyClient.sendUseItemOnNpc(selectedInventorySlot, (int) pickedEntityId);
                     }
                     selectedInventorySlot = -1;
                     sidePanel.setSelectedInventorySlot(-1);
